New buyer registration pages are ready
the buyer registration page will just take the industry details and the preferred password for their user account. The userid of the buyer will be generated by the server itself based on the details the applicant provided. ImageDisplayHelperServlet.java is modified. Two emails are sent in each registration process, one sends the unique 6-digit code to the applicant provided email address, and once the email is verified to be valid, then the server-generated userid for the buyer will be sent to the valid email address. once the userid is generated and a successful registration message is displayed to the user, the user is redirected to the login page.